+++ test/test_apply_function_pullbacks.py
@@ -3,7 +3,7 @@ import numpy as np
 from ufl import Cell, Coefficient, FunctionSpace, Mesh, as_tensor, as_vector, dx, indices, triangle
 from ufl.algorithms.renumbering import renumber_indices
 from ufl.classes import Jacobian, JacobianDeterminant, JacobianInverse, ReferenceValue
-from ufl.finiteelement import FiniteElement, MixedElement, SymmetricElement
+from utils import FiniteElement, MixedElement, SymmetricElement
 from ufl.pullback import (
     contravariant_piola,
     covariant_piola,
