+++ test/test_arithmetic.py
@@ -13,7 +13,7 @@ from ufl import (
     triangle,
 )
 from ufl.classes import ComplexValue, Division, FloatValue, IntValue
-from ufl.finiteelement import FiniteElement
+from utils import FiniteElement
 from ufl.pullback import identity_pullback
 from ufl.sobolevspace import H1
