+++ test/test_complex.py
@@ -41,7 +41,7 @@ from ufl.algorithms.comparison_checker import ComplexComparisonError, do_compari
 from ufl.algorithms.formtransformations import compute_form_adjoint
 from ufl.algorithms.remove_complex_nodes import remove_complex_nodes
 from ufl.constantvalue import ComplexValue, Zero
-from ufl.finiteelement import FiniteElement
+from utils import FiniteElement
 from ufl.pullback import identity_pullback
 from ufl.sobolevspace import H1
