+++ test/test_equals.py
@@ -2,7 +2,7 @@
 
 from ufl import Coefficient, Cofunction, FunctionSpace, Mesh, triangle
 from ufl.exprcontainers import ExprList
-from ufl.finiteelement import FiniteElement
+from utils import FiniteElement
 from ufl.pullback import identity_pullback
 from ufl.sobolevspace import H1
