+++ test/test_external_operator.py
@@ -30,7 +30,7 @@ from ufl import (
 from ufl.algorithms import expand_derivatives
 from ufl.algorithms.apply_derivatives import apply_derivatives
 from ufl.core.external_operator import ExternalOperator
-from ufl.finiteelement import FiniteElement
+from utils import FiniteElement
 from ufl.form import BaseForm, ZeroBaseForm
 from ufl.pullback import identity_pullback
 from ufl.sobolevspace import H1
