+++ test/test_ffcforms.py
@@ -43,7 +43,7 @@ from ufl import (
     tetrahedron,
     triangle,
 )
-from ufl.finiteelement import FiniteElement, MixedElement
+from utils import FiniteElement, MixedElement
 from ufl.pullback import contravariant_piola, covariant_piola, identity_pullback
 from ufl.sobolevspace import H1, L2, HCurl, HDiv
