+++ test/test_form.py
@@ -20,7 +20,7 @@ from ufl import (
     nabla_grad,
     triangle,
 )
-from ufl.finiteelement import FiniteElement
+from utils import FiniteElement
 from ufl.form import BaseForm
 from ufl.pullback import identity_pullback
 from ufl.sobolevspace import H1
