+++ test/test_illegal.py
@@ -1,7 +1,7 @@
 import pytest
 
 from ufl import Argument, Coefficient, FunctionSpace, Mesh, triangle
-from ufl.finiteelement import FiniteElement
+from utils import FiniteElement
 from ufl.pullback import identity_pullback
 from ufl.sobolevspace import H1
