+++ test/test_indices.py
@@ -26,7 +26,7 @@ from ufl import (
     triangle,
 )
 from ufl.classes import IndexSum
-from ufl.finiteelement import FiniteElement
+from utils import FiniteElement
 from ufl.pullback import identity_pullback
 from ufl.sobolevspace import H1
