+++ test/test_measures.py
@@ -3,7 +3,7 @@
 from mockobjects import MockMesh, MockMeshFunction
 
 from ufl import Cell, Coefficient, FunctionSpace, Measure, Mesh, as_ufl, dC, dI, dO, triangle
-from ufl.finiteelement import FiniteElement
+from utils import FiniteElement
 from ufl.pullback import identity_pullback
 from ufl.sobolevspace import H1
