+++ test/test_pickle.py
@@ -42,7 +42,7 @@ from ufl import (
     triangle,
 )
 from ufl.algorithms import compute_form_data
-from ufl.finiteelement import FiniteElement, MixedElement
+from utils import FiniteElement, MixedElement
 from ufl.pullback import contravariant_piola, covariant_piola, identity_pullback
 from ufl.sobolevspace import H1, L2, HCurl, HDiv
