+++ test/test_scratch.py
@@ -25,7 +25,7 @@ from ufl import (
     triangle,
 )
 from ufl.classes import FixedIndex, FormArgument, Grad, Indexed, ListTensor, Zero
-from ufl.finiteelement import FiniteElement
+from utils import FiniteElement
 from ufl.pullback import identity_pullback
 from ufl.sobolevspace import H1
 from ufl.tensors import as_scalar, unit_indexed_tensor, unwrap_list_tensor
